♻️REFACTOR : MockList 생성, QueueTests 수정
