--- BankManagerConsoleApp/QueueTests/QueueTests.swift
@@ -1,53 +1,0 @@
-//
-//  LinkedListTests.swift
-//  LinkedListTests
-//
-//  Created by OneTool, Red on 2022/04/26.
-//
-
-import XCTest
-
-class QueueTests: XCTestCase {
-    var sut: Queue<LinkedList<Int>>!
-    
-    override func setUp() {
-       sut = Queue(listType: LinkedList<Int>())
-    }
-
-    override func tearDown() {
-        self.sut = nil
-    }
-
-    func test_isEmpty() {
-        sut.enQueue(data: 1)
-        sut.enQueue(data: 2)
-        sut.enQueue(data: 3)
-
-        sut.clear()
-        let result = sut.isEmpty
-
-        XCTAssertTrue(result)
-    }
-
-    func test_peek() {
-        sut.enQueue(data: 1)
-        sut.enQueue(data: 2)
-        sut.enQueue(data: 3)
-
-        let result = sut.peek
-
-        XCTAssertEqual(result, 1)
-    }
-
-    func test_deQueue() {
-        sut.enQueue(data: 1)
-        sut.enQueue(data: 2)
-        sut.enQueue(data: 3)
-
-        let result1 = sut.deQueue()
-        let result2 = sut.peek
-
-        XCTAssertEqual(result1, 1)
-        XCTAssertEqual(result2, 2)
-    }
-}
